requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.2
streamlit-echarts>=0.4.0
streamlit-option-menu>=0.3.6
openai>=1.3.5
supabase>=1.2.0
httpx>=0.26.0
python-dateutil>=2.8.2